Export `inspectYText`
