+++ packages/core/src/utils/constants.ts
@@ -1,0 +1,7 @@
+export const STORED_POSITION_PREFIX = '__slateYjsStoredPosition_';
+
+/**
+ * Represent empty text nodes using a character with a special attribute, since
+ * Yjs doesn't support storing attributes on empty insertions.
+ */
+export const EMPTY_TEXT_PREFIX = '__slateYjsEmptyText_';
