+++ packages/core/test/utils/ClonedSharedRoot.test.ts
@@ -1,7 +1,7 @@
 import * as Y from 'yjs';
 import { describe, it, expect, beforeEach } from 'vitest';
-import { inspectYText } from '../support/utils';
 import { ClonedSharedRoot } from '../../src/utils/ClonedSharedRoot';
+import { inspectYText } from '../../src/utils/yjs';
 
 describe('ClonedSharedRoot', () => {
   let doc: Y.Doc;
